refactor: Remove all instances of FindIter in a non-sorted case and move from submodule back into individual file
