+++ src/types/iterator/mod.rs
@@ -7,7 +7,7 @@ mod merge;
 mod subtract;
 pub use cluster::ClusterIter;
 pub use complement::ComplementIter;
-pub use find::{FindIter, FindIterOwned, FindIterSorted, FindIterSortedOwned};
+pub use find::{FindIter, FindIterOwned};
 pub use intersect::IntersectIter;
 pub use interval::{IntervalIterOwned, IntervalIterRef};
 pub use merge::MergeIter;
